activity-agent/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
python-dotenv>=0.19.0
httpx>=0.23.0
python-multipart>=0.0.5
requests>=2.26.0 